#3843 table separate component + api/hook
