+++ src/frontend/src/pages/RulesPage/RulesetTypePage.tsx
@@ -1,147 +1,13 @@
 import PageLayout from '../../components/PageLayout';
-import {
-  Box,
-  Button,
-  Paper,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  useMediaQuery,
-  useTheme,
-  Card,
-  CardContent,
-  Typography,
-  Stack
-} from '@mui/material';
-import { datePipe } from '../../utils/pipes';
-import { NERButton } from '../../components/NERButton';
-import { useHistory } from 'react-router-dom';
-import { routes } from '../../utils/routes';
-
-type RulesetTypeColumnId = 'id' | 'name' | 'lastUpdated' | 'revisions' | 'actions';
-
-interface RulesetTypeHeadCell {
-  id: RulesetTypeColumnId;
-  label: string;
-}
+import { Box, Button } from '@mui/material';
+import RulesetTypeTable from './components/RulesetTypeTable';
 
 const RulesetTypePage: React.FC = () => {
-  const history = useHistory();
-  const theme = useTheme();
-  const isMobile = useMediaQuery(theme.breakpoints.down('md'));
-
-  const headCells: readonly RulesetTypeHeadCell[] = [
-    {
-      id: 'name',
-      label: 'Ruleset Name'
-    },
-    {
-      id: 'lastUpdated',
-      label: 'Last Updated'
-    },
-    {
-      id: 'revisions',
-      label: 'Number of Revisions'
-    },
-    {
-      id: 'actions',
-      label: 'Actions'
-    }
-  ];
-
-  // Mock data for now - will be replaced with ruleset type data
-  const mockRulesetTypes = [
-    { id: '1', name: 'Ruleset 1', lastUpdated: new Date('2024-01-15'), revisions: 5, actions: 'Edit' },
-    { id: '2', name: 'Ruleset 2', lastUpdated: new Date('2024-01-14'), revisions: 3, actions: 'Edit' }
-  ];
-
   return (
     <PageLayout title="Ruleset Types">
       <Box sx={{ display: 'flex', flexDirection: 'column', minHeight: 'calc(100vh - 120px)' }}>
         <Box sx={{ flexGrow: 1 }}>
-          {isMobile ? (
-            <Stack spacing={2} sx={{ px: 1 }}>
-              {mockRulesetTypes.map((rulesetType) => (
-                <Card
-                  key={rulesetType.id}
-                  sx={{
-                    backgroundColor: '#121313',
-                    borderRadius: '8px',
-                    border: '1px solid rgba(255, 255, 255, 0.1)'
-                  }}
-                >
-                  <CardContent>
-                    <Typography variant="h6" sx={{ color: '#dd514c', fontWeight: 600, mb: 2 }}>
-                      {rulesetType.name}
-                    </Typography>
-                    <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
-                      <Box sx={{ display: 'flex', justifyContent: 'space-between' }}>
-                        <Typography variant="body2" sx={{ color: '#999' }}>
-                          Last Updated:
-                        </Typography>
-                        <Typography variant="body2" sx={{ color: '#ededed' }}>
-                          {datePipe(rulesetType.lastUpdated)}
-                        </Typography>
-                      </Box>
-                      <Box sx={{ display: 'flex', justifyContent: 'space-between' }}>
-                        <Typography variant="body2" sx={{ color: '#999' }}>
-                          Revisions:
-                        </Typography>
-                        <Typography variant="body2" sx={{ color: '#ededed' }}>
-                          {rulesetType.revisions}
-                        </Typography>
-                      </Box>
-                      <Box sx={{ display: 'flex', justifyContent: 'flex-end', mt: 1 }}>
-                        <Button
-                          size="small"
-                          variant="outlined"
-                          sx={{
-                            color: '#dd514c',
-                            borderColor: '#dd514c',
-                            '&:hover': {
-                              borderColor: '#c74340',
-                              backgroundColor: 'rgba(221, 81, 76, 0.08)'
-                            }
-                          }}
-                        >
-                          Actions
-                        </Button>
-                      </Box>
-                    </Box>
-                  </CardContent>
-                </Card>
-              ))}
-            </Stack>
-          ) : (
-            <TableContainer component={Paper} sx={{ borderRadius: '8px', overflow: 'hidden' }}>
-              <Table aria-label="ruleset types">
-                <TableHead
-                  sx={{
-                    backgroundColor: '#dd514c'
-                  }}
-                >
-                  <TableRow>
-                    {headCells.map((headCell) => (
-                      <TableCell
-                        align="center"
-                        sx={{ fontSize: '16px', fontWeight: 600 }}
-                        style={{ paddingLeft: '24px', paddingRight: '0px' }}
-                        key={headCell.id}
-                      >
-                        {headCell.label}
-                      </TableCell>
-                    ))}
-                  </TableRow>
-                </TableHead>
-                <TableBody sx={{ backgroundColor: '#121313' }}>
-                  {/* Table rows will go here with ruleset type data */}
-                </TableBody>
-              </Table>
-            </TableContainer>
-          )}
+          <RulesetTypeTable />
         </Box>
         <Box
           sx={{
@@ -188,8 +54,6 @@ const RulesetTypePage: React.FC = () => {
             >
               Add Ruleset
             </Button>
-            {/* Temporary for navigation */}
-            <NERButton onClick={() => history.push(`${routes.RULES}/placeholder_ruleset_id`)}>FSAE Ruleset</NERButton>
           </Box>
         </Box>
       </Box>
